% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_rplaywright.R
\name{add_init_script}
\alias{add_init_script}
\title{Adds a script which would be evaluated in one of the following scenarios:}
\usage{
add_init_script(page, ...)
}
\arguments{
\item{page}{Page instance}
}
\description{
Adds a script which would be evaluated in one of the following scenarios:
}
\examples{
\dontrun{
 page <- rplaywright::new_chromium() \%>\%
  new_context() \%>\%
  new_page()

 page \%>\% add_init_script(list(path=normalizePath("./examples/preload.js")))
}
}
\references{
\itemize{
\item \link{https://erikaris.github.io/rplaywright/#add_init_script}
}
}
