% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_browser.R
\name{Browser}
\alias{Browser}
\title{Browser Class}
\description{
Browser Class

Browser Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Browser-close}{\code{Browser$close()}}
\item \href{#method-Browser-new_context}{\code{Browser$new_context()}}
\item \href{#method-Browser-new}{\code{Browser$new()}}
\item \href{#method-Browser-launch}{\code{Browser$launch()}}
\item \href{#method-Browser-clone}{\code{Browser$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Browser-close"></a>}}
\if{latex}{\out{\hypertarget{method-Browser-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Browser$close(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Browser-new_context"></a>}}
\if{latex}{\out{\hypertarget{method-Browser-new_context}{}}}
\subsection{Method \code{new_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Browser$new_context(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Browser-new"></a>}}
\if{latex}{\out{\hypertarget{method-Browser-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Browser$new(type, remote_url = "http://localhost:3000")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Browser-launch"></a>}}
\if{latex}{\out{\hypertarget{method-Browser-launch}{}}}
\subsection{Method \code{launch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Browser$launch()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Browser-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Browser-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Browser$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
