% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_context.R
\name{Context}
\alias{Context}
\title{Context Class}
\description{
Context Class

Context Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Context-new_page}{\code{Context$new_page()}}
\item \href{#method-Context-new}{\code{Context$new()}}
\item \href{#method-Context-clone}{\code{Context$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Context-new_page"></a>}}
\if{latex}{\out{\hypertarget{method-Context-new_page}{}}}
\subsection{Method \code{new_page()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Context$new_page(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Context-new"></a>}}
\if{latex}{\out{\hypertarget{method-Context-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Context$new(parent, options = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Context-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Context-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Context$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
