% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_frame.R
\name{Frame}
\alias{Frame}
\title{Frame Class}
\description{
Frame Class

Frame Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Frame-frame_element}{\code{Frame$frame_element()}}
\item \href{#method-Frame-add_script_tag}{\code{Frame$add_script_tag()}}
\item \href{#method-Frame-add_style_tag}{\code{Frame$add_style_tag()}}
\item \href{#method-Frame-child_frames}{\code{Frame$child_frames()}}
\item \href{#method-Frame-content}{\code{Frame$content()}}
\item \href{#method-Frame-drag_and_drop}{\code{Frame$drag_and_drop()}}
\item \href{#method-Frame-evaluate}{\code{Frame$evaluate()}}
\item \href{#method-Frame-evaluate_handle}{\code{Frame$evaluate_handle()}}
\item \href{#method-Frame-frame_locator}{\code{Frame$frame_locator()}}
\item \href{#method-Frame-get_by_alt_text}{\code{Frame$get_by_alt_text()}}
\item \href{#method-Frame-get_by_label}{\code{Frame$get_by_label()}}
\item \href{#method-Frame-get_by_placeholder}{\code{Frame$get_by_placeholder()}}
\item \href{#method-Frame-get_by_role}{\code{Frame$get_by_role()}}
\item \href{#method-Frame-get_by_test_id}{\code{Frame$get_by_test_id()}}
\item \href{#method-Frame-get_by_text}{\code{Frame$get_by_text()}}
\item \href{#method-Frame-get_by_title}{\code{Frame$get_by_title()}}
\item \href{#method-Frame-goto}{\code{Frame$goto()}}
\item \href{#method-Frame-is_detached}{\code{Frame$is_detached()}}
\item \href{#method-Frame-is_enabled}{\code{Frame$is_enabled()}}
\item \href{#method-Frame-locator}{\code{Frame$locator()}}
\item \href{#method-Frame-name}{\code{Frame$name()}}
\item \href{#method-Frame-page}{\code{Frame$page()}}
\item \href{#method-Frame-parent_frame}{\code{Frame$parent_frame()}}
\item \href{#method-Frame-set_content}{\code{Frame$set_content()}}
\item \href{#method-Frame-title}{\code{Frame$title()}}
\item \href{#method-Frame-url}{\code{Frame$url()}}
\item \href{#method-Frame-wait_for_function}{\code{Frame$wait_for_function()}}
\item \href{#method-Frame-wait_for_load_state}{\code{Frame$wait_for_load_state()}}
\item \href{#method-Frame-wait_for_u_r_l}{\code{Frame$wait_for_u_r_l()}}
\item \href{#method-Frame-new}{\code{Frame$new()}}
\item \href{#method-Frame-clone}{\code{Frame$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-frame_element"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-frame_element}{}}}
\subsection{Method \code{frame_element()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$frame_element()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-add_script_tag"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-add_script_tag}{}}}
\subsection{Method \code{add_script_tag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$add_script_tag(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-add_style_tag"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-add_style_tag}{}}}
\subsection{Method \code{add_style_tag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$add_style_tag(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-child_frames"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-child_frames}{}}}
\subsection{Method \code{child_frames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$child_frames(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-content"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-content}{}}}
\subsection{Method \code{content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$content(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-drag_and_drop"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-drag_and_drop}{}}}
\subsection{Method \code{drag_and_drop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$drag_and_drop(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$evaluate(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-evaluate_handle"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-evaluate_handle}{}}}
\subsection{Method \code{evaluate_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$evaluate_handle(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-frame_locator"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-frame_locator}{}}}
\subsection{Method \code{frame_locator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$frame_locator(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-get_by_alt_text"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-get_by_alt_text}{}}}
\subsection{Method \code{get_by_alt_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$get_by_alt_text(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-get_by_label"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-get_by_label}{}}}
\subsection{Method \code{get_by_label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$get_by_label(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-get_by_placeholder"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-get_by_placeholder}{}}}
\subsection{Method \code{get_by_placeholder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$get_by_placeholder(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-get_by_role"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-get_by_role}{}}}
\subsection{Method \code{get_by_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$get_by_role(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-get_by_test_id"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-get_by_test_id}{}}}
\subsection{Method \code{get_by_test_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$get_by_test_id(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-get_by_text"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-get_by_text}{}}}
\subsection{Method \code{get_by_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$get_by_text(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-get_by_title"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-get_by_title}{}}}
\subsection{Method \code{get_by_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$get_by_title(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-goto"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-goto}{}}}
\subsection{Method \code{goto()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$goto(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-is_detached"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-is_detached}{}}}
\subsection{Method \code{is_detached()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$is_detached(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-is_enabled"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-is_enabled}{}}}
\subsection{Method \code{is_enabled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$is_enabled(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-locator"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-locator}{}}}
\subsection{Method \code{locator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$locator(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-name"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$name(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-page"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-page}{}}}
\subsection{Method \code{page()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$page(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-parent_frame"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-parent_frame}{}}}
\subsection{Method \code{parent_frame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$parent_frame(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-set_content"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-set_content}{}}}
\subsection{Method \code{set_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$set_content(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-title"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-title}{}}}
\subsection{Method \code{title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$title(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-url"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-url}{}}}
\subsection{Method \code{url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$url(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-wait_for_function"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-wait_for_function}{}}}
\subsection{Method \code{wait_for_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$wait_for_function(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-wait_for_load_state"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-wait_for_load_state}{}}}
\subsection{Method \code{wait_for_load_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$wait_for_load_state(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-wait_for_u_r_l"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-wait_for_u_r_l}{}}}
\subsection{Method \code{wait_for_u_r_l()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$wait_for_u_r_l(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-new"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Frame-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Frame-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Frame$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
