% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_frame_locator.R
\name{FrameLocator}
\alias{FrameLocator}
\title{FrameLocator Class}
\description{
FrameLocator Class

FrameLocator Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FrameLocator-first}{\code{FrameLocator$first()}}
\item \href{#method-FrameLocator-frame_locator}{\code{FrameLocator$frame_locator()}}
\item \href{#method-FrameLocator-get_by_alt_text}{\code{FrameLocator$get_by_alt_text()}}
\item \href{#method-FrameLocator-get_by_label}{\code{FrameLocator$get_by_label()}}
\item \href{#method-FrameLocator-get_by_placeholder}{\code{FrameLocator$get_by_placeholder()}}
\item \href{#method-FrameLocator-get_by_role}{\code{FrameLocator$get_by_role()}}
\item \href{#method-FrameLocator-get_by_test_id}{\code{FrameLocator$get_by_test_id()}}
\item \href{#method-FrameLocator-get_by_text}{\code{FrameLocator$get_by_text()}}
\item \href{#method-FrameLocator-get_by_title}{\code{FrameLocator$get_by_title()}}
\item \href{#method-FrameLocator-last}{\code{FrameLocator$last()}}
\item \href{#method-FrameLocator-locator}{\code{FrameLocator$locator()}}
\item \href{#method-FrameLocator-nth}{\code{FrameLocator$nth()}}
\item \href{#method-FrameLocator-owner}{\code{FrameLocator$owner()}}
\item \href{#method-FrameLocator-new}{\code{FrameLocator$new()}}
\item \href{#method-FrameLocator-clone}{\code{FrameLocator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-first"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-first}{}}}
\subsection{Method \code{first()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$first(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-frame_locator"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-frame_locator}{}}}
\subsection{Method \code{frame_locator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$frame_locator(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-get_by_alt_text"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-get_by_alt_text}{}}}
\subsection{Method \code{get_by_alt_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$get_by_alt_text(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-get_by_label"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-get_by_label}{}}}
\subsection{Method \code{get_by_label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$get_by_label(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-get_by_placeholder"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-get_by_placeholder}{}}}
\subsection{Method \code{get_by_placeholder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$get_by_placeholder(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-get_by_role"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-get_by_role}{}}}
\subsection{Method \code{get_by_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$get_by_role(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-get_by_test_id"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-get_by_test_id}{}}}
\subsection{Method \code{get_by_test_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$get_by_test_id(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-get_by_text"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-get_by_text}{}}}
\subsection{Method \code{get_by_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$get_by_text(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-get_by_title"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-get_by_title}{}}}
\subsection{Method \code{get_by_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$get_by_title(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-last"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-last}{}}}
\subsection{Method \code{last()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$last(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-locator"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-locator}{}}}
\subsection{Method \code{locator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$locator(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-nth"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-nth}{}}}
\subsection{Method \code{nth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$nth(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-owner"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-owner}{}}}
\subsection{Method \code{owner()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$owner(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-new"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FrameLocator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FrameLocator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FrameLocator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
