% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_install.R
\name{install_rplaywright}
\alias{install_rplaywright}
\title{Install rplaywright and its dependencies}
\usage{
install_rplaywright(force = FALSE)
}
\arguments{
\item{force}{Install dependencies without prompt}
}
\description{
Install rplaywright and its dependencies
}
\examples{
\dontrun{
  rplaywright::install_rplaywright(force=TRUE)
}

}
