% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_jshandle.R
\name{JSHandle}
\alias{JSHandle}
\title{JSHandle Class}
\description{
JSHandle Class

JSHandle Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JSHandle-as_element}{\code{JSHandle$as_element()}}
\item \href{#method-JSHandle-dispose}{\code{JSHandle$dispose()}}
\item \href{#method-JSHandle-evaluate}{\code{JSHandle$evaluate()}}
\item \href{#method-JSHandle-evaluate_handle}{\code{JSHandle$evaluate_handle()}}
\item \href{#method-JSHandle-get_properties}{\code{JSHandle$get_properties()}}
\item \href{#method-JSHandle-get_property}{\code{JSHandle$get_property()}}
\item \href{#method-JSHandle-json_value}{\code{JSHandle$json_value()}}
\item \href{#method-JSHandle-new}{\code{JSHandle$new()}}
\item \href{#method-JSHandle-clone}{\code{JSHandle$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSHandle-as_element"></a>}}
\if{latex}{\out{\hypertarget{method-JSHandle-as_element}{}}}
\subsection{Method \code{as_element()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSHandle$as_element(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSHandle-dispose"></a>}}
\if{latex}{\out{\hypertarget{method-JSHandle-dispose}{}}}
\subsection{Method \code{dispose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSHandle$dispose(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSHandle-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-JSHandle-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSHandle$evaluate(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSHandle-evaluate_handle"></a>}}
\if{latex}{\out{\hypertarget{method-JSHandle-evaluate_handle}{}}}
\subsection{Method \code{evaluate_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSHandle$evaluate_handle(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSHandle-get_properties"></a>}}
\if{latex}{\out{\hypertarget{method-JSHandle-get_properties}{}}}
\subsection{Method \code{get_properties()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSHandle$get_properties(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSHandle-get_property"></a>}}
\if{latex}{\out{\hypertarget{method-JSHandle-get_property}{}}}
\subsection{Method \code{get_property()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSHandle$get_property(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSHandle-json_value"></a>}}
\if{latex}{\out{\hypertarget{method-JSHandle-json_value}{}}}
\subsection{Method \code{json_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSHandle$json_value(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSHandle-new"></a>}}
\if{latex}{\out{\hypertarget{method-JSHandle-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSHandle$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JSHandle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JSHandle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSHandle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
