% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_locator.R
\name{Locator}
\alias{Locator}
\title{Locator Class}
\description{
Locator Class

Locator Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Locator-and}{\code{Locator$and()}}
\item \href{#method-Locator-content_frame}{\code{Locator$content_frame()}}
\item \href{#method-Locator-drag_to}{\code{Locator$drag_to()}}
\item \href{#method-Locator-evaluate}{\code{Locator$evaluate()}}
\item \href{#method-Locator-evaluate_all}{\code{Locator$evaluate_all()}}
\item \href{#method-Locator-evaluate_handle}{\code{Locator$evaluate_handle()}}
\item \href{#method-Locator-filter}{\code{Locator$filter()}}
\item \href{#method-Locator-frame_locator}{\code{Locator$frame_locator()}}
\item \href{#method-Locator-or}{\code{Locator$or()}}
\item \href{#method-Locator-page}{\code{Locator$page()}}
\item \href{#method-Locator-set_input_files}{\code{Locator$set_input_files()}}
\item \href{#method-Locator-tap}{\code{Locator$tap()}}
\item \href{#method-Locator-all}{\code{Locator$all()}}
\item \href{#method-Locator-all_inner_texts}{\code{Locator$all_inner_texts()}}
\item \href{#method-Locator-all_text_contents}{\code{Locator$all_text_contents()}}
\item \href{#method-Locator-blur}{\code{Locator$blur()}}
\item \href{#method-Locator-bounding_box}{\code{Locator$bounding_box()}}
\item \href{#method-Locator-check}{\code{Locator$check()}}
\item \href{#method-Locator-clear}{\code{Locator$clear()}}
\item \href{#method-Locator-click}{\code{Locator$click()}}
\item \href{#method-Locator-count}{\code{Locator$count()}}
\item \href{#method-Locator-dblclick}{\code{Locator$dblclick()}}
\item \href{#method-Locator-dispatch_event}{\code{Locator$dispatch_event()}}
\item \href{#method-Locator-fill}{\code{Locator$fill()}}
\item \href{#method-Locator-first}{\code{Locator$first()}}
\item \href{#method-Locator-focus}{\code{Locator$focus()}}
\item \href{#method-Locator-get_attribute}{\code{Locator$get_attribute()}}
\item \href{#method-Locator-get_by_alt_text}{\code{Locator$get_by_alt_text()}}
\item \href{#method-Locator-get_by_label}{\code{Locator$get_by_label()}}
\item \href{#method-Locator-get_by_placeholder}{\code{Locator$get_by_placeholder()}}
\item \href{#method-Locator-get_by_role}{\code{Locator$get_by_role()}}
\item \href{#method-Locator-get_by_test_id}{\code{Locator$get_by_test_id()}}
\item \href{#method-Locator-get_by_text}{\code{Locator$get_by_text()}}
\item \href{#method-Locator-get_by_title}{\code{Locator$get_by_title()}}
\item \href{#method-Locator-highlight}{\code{Locator$highlight()}}
\item \href{#method-Locator-hover}{\code{Locator$hover()}}
\item \href{#method-Locator-inner_h_t_m_l}{\code{Locator$inner_h_t_m_l()}}
\item \href{#method-Locator-inner_text}{\code{Locator$inner_text()}}
\item \href{#method-Locator-input_value}{\code{Locator$input_value()}}
\item \href{#method-Locator-is_checked}{\code{Locator$is_checked()}}
\item \href{#method-Locator-is_disabled}{\code{Locator$is_disabled()}}
\item \href{#method-Locator-is_editable}{\code{Locator$is_editable()}}
\item \href{#method-Locator-is_enabled}{\code{Locator$is_enabled()}}
\item \href{#method-Locator-is_hidden}{\code{Locator$is_hidden()}}
\item \href{#method-Locator-is_visible}{\code{Locator$is_visible()}}
\item \href{#method-Locator-last}{\code{Locator$last()}}
\item \href{#method-Locator-locator}{\code{Locator$locator()}}
\item \href{#method-Locator-nth}{\code{Locator$nth()}}
\item \href{#method-Locator-parent}{\code{Locator$parent()}}
\item \href{#method-Locator-press}{\code{Locator$press()}}
\item \href{#method-Locator-press_sequentially}{\code{Locator$press_sequentially()}}
\item \href{#method-Locator-screenshot}{\code{Locator$screenshot()}}
\item \href{#method-Locator-scroll_into_view_if_needed}{\code{Locator$scroll_into_view_if_needed()}}
\item \href{#method-Locator-select_option}{\code{Locator$select_option()}}
\item \href{#method-Locator-select_text}{\code{Locator$select_text()}}
\item \href{#method-Locator-set_checked}{\code{Locator$set_checked()}}
\item \href{#method-Locator-text_content}{\code{Locator$text_content()}}
\item \href{#method-Locator-uncheck}{\code{Locator$uncheck()}}
\item \href{#method-Locator-wait_for}{\code{Locator$wait_for()}}
\item \href{#method-Locator-new}{\code{Locator$new()}}
\item \href{#method-Locator-clone}{\code{Locator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-and"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-and}{}}}
\subsection{Method \code{and()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$and()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-content_frame"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-content_frame}{}}}
\subsection{Method \code{content_frame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$content_frame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-drag_to"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-drag_to}{}}}
\subsection{Method \code{drag_to()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$drag_to()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$evaluate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-evaluate_all"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-evaluate_all}{}}}
\subsection{Method \code{evaluate_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$evaluate_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-evaluate_handle"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-evaluate_handle}{}}}
\subsection{Method \code{evaluate_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$evaluate_handle()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-filter"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-filter}{}}}
\subsection{Method \code{filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$filter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-frame_locator"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-frame_locator}{}}}
\subsection{Method \code{frame_locator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$frame_locator()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-or"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-or}{}}}
\subsection{Method \code{or()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$or()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-page"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-page}{}}}
\subsection{Method \code{page()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$page()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-set_input_files"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-set_input_files}{}}}
\subsection{Method \code{set_input_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$set_input_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-tap"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-tap}{}}}
\subsection{Method \code{tap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$tap()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-all"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-all}{}}}
\subsection{Method \code{all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$all(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-all_inner_texts"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-all_inner_texts}{}}}
\subsection{Method \code{all_inner_texts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$all_inner_texts(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-all_text_contents"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-all_text_contents}{}}}
\subsection{Method \code{all_text_contents()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$all_text_contents(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-blur"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-blur}{}}}
\subsection{Method \code{blur()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$blur(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-bounding_box"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-bounding_box}{}}}
\subsection{Method \code{bounding_box()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$bounding_box(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-check"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-check}{}}}
\subsection{Method \code{check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$check(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-clear"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-clear}{}}}
\subsection{Method \code{clear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$clear(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-click"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-click}{}}}
\subsection{Method \code{click()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$click(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-count"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-count}{}}}
\subsection{Method \code{count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$count(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-dblclick"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-dblclick}{}}}
\subsection{Method \code{dblclick()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$dblclick(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-dispatch_event"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-dispatch_event}{}}}
\subsection{Method \code{dispatch_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$dispatch_event(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-fill"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-fill}{}}}
\subsection{Method \code{fill()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$fill(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-first"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-first}{}}}
\subsection{Method \code{first()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$first(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-focus"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-focus}{}}}
\subsection{Method \code{focus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$focus(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-get_attribute"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-get_attribute}{}}}
\subsection{Method \code{get_attribute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$get_attribute(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-get_by_alt_text"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-get_by_alt_text}{}}}
\subsection{Method \code{get_by_alt_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$get_by_alt_text(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-get_by_label"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-get_by_label}{}}}
\subsection{Method \code{get_by_label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$get_by_label(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-get_by_placeholder"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-get_by_placeholder}{}}}
\subsection{Method \code{get_by_placeholder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$get_by_placeholder(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-get_by_role"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-get_by_role}{}}}
\subsection{Method \code{get_by_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$get_by_role(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-get_by_test_id"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-get_by_test_id}{}}}
\subsection{Method \code{get_by_test_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$get_by_test_id(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-get_by_text"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-get_by_text}{}}}
\subsection{Method \code{get_by_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$get_by_text(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-get_by_title"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-get_by_title}{}}}
\subsection{Method \code{get_by_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$get_by_title(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-highlight"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-highlight}{}}}
\subsection{Method \code{highlight()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$highlight(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-hover"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-hover}{}}}
\subsection{Method \code{hover()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$hover(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-inner_h_t_m_l"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-inner_h_t_m_l}{}}}
\subsection{Method \code{inner_h_t_m_l()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$inner_h_t_m_l(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-inner_text"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-inner_text}{}}}
\subsection{Method \code{inner_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$inner_text(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-input_value"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-input_value}{}}}
\subsection{Method \code{input_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$input_value(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-is_checked"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-is_checked}{}}}
\subsection{Method \code{is_checked()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$is_checked(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-is_disabled"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-is_disabled}{}}}
\subsection{Method \code{is_disabled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$is_disabled(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-is_editable"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-is_editable}{}}}
\subsection{Method \code{is_editable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$is_editable(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-is_enabled"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-is_enabled}{}}}
\subsection{Method \code{is_enabled()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$is_enabled(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-is_hidden"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-is_hidden}{}}}
\subsection{Method \code{is_hidden()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$is_hidden(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-is_visible"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-is_visible}{}}}
\subsection{Method \code{is_visible()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$is_visible(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-last"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-last}{}}}
\subsection{Method \code{last()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$last(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-locator"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-locator}{}}}
\subsection{Method \code{locator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$locator(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-nth"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-nth}{}}}
\subsection{Method \code{nth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$nth(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-parent"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-parent}{}}}
\subsection{Method \code{parent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$parent(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-press"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-press}{}}}
\subsection{Method \code{press()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$press(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-press_sequentially"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-press_sequentially}{}}}
\subsection{Method \code{press_sequentially()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$press_sequentially(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-screenshot"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-screenshot}{}}}
\subsection{Method \code{screenshot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$screenshot(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-scroll_into_view_if_needed"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-scroll_into_view_if_needed}{}}}
\subsection{Method \code{scroll_into_view_if_needed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$scroll_into_view_if_needed(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-select_option"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-select_option}{}}}
\subsection{Method \code{select_option()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$select_option(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-select_text"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-select_text}{}}}
\subsection{Method \code{select_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$select_text(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-set_checked"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-set_checked}{}}}
\subsection{Method \code{set_checked()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$set_checked(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-text_content"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-text_content}{}}}
\subsection{Method \code{text_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$text_content(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-uncheck"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-uncheck}{}}}
\subsection{Method \code{uncheck()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$uncheck(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-wait_for"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-wait_for}{}}}
\subsection{Method \code{wait_for()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$wait_for(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Locator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Locator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Locator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
