% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_rplaywright.R
\name{new_context}
\alias{new_context}
\title{Creates a new browser context. It won't share cookies/cache with other browser contexts.}
\usage{
new_context(browser)
}
\arguments{
\item{browser}{Browser instance}
}
\description{
Creates a new browser context. It won't share cookies/cache with other browser contexts.
}
\examples{
\dontrun{
 rplaywright::new_chromium() \%>\%
  new_context()
}
}
