% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_rplaywright.R
\name{new_firefox}
\alias{new_firefox}
\title{Launch new firefox instance}
\usage{
new_firefox(
  start_server = T,
  serverOptions = list(host = "127.0.0.1", port = 3000)
)
}
\description{
Launch new firefox instance
}
\examples{
browser <- rplaywright::new_firefox()

}
