% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_rplaywright.R
\name{new_page}
\alias{new_page}
\title{Creates a new page in the browser context.}
\usage{
new_page(context)
}
\arguments{
\item{context}{Browser Context instance}
}
\description{
Creates a new page in the browser context.
}
\examples{
\dontrun{
 rplaywright::new_chromium() \%>\%
  new_context() \%>\%
  new_page()
}
}
