% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_rplaywright.R
\name{new_webkit}
\alias{new_webkit}
\title{Launch new webkit instance}
\usage{
new_webkit(start_server = F)
}
\description{
Launch new webkit instance
}
\examples{
browser <- rplaywright::new_webkit()

}
