% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_page.R
\name{Page}
\alias{Page}
\title{Page Class}
\description{
Page Class

Page Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Page-add_locator_handler}{\code{Page$add_locator_handler()}}
\item \href{#method-Page-remove_locator_handler}{\code{Page$remove_locator_handler()}}
\item \href{#method-Page-route_from_h_a_r}{\code{Page$route_from_h_a_r()}}
\item \href{#method-Page-add_init_script}{\code{Page$add_init_script()}}
\item \href{#method-Page-add_script_tag}{\code{Page$add_script_tag()}}
\item \href{#method-Page-add_style_tag}{\code{Page$add_style_tag()}}
\item \href{#method-Page-bring_to_front}{\code{Page$bring_to_front()}}
\item \href{#method-Page-close}{\code{Page$close()}}
\item \href{#method-Page-content}{\code{Page$content()}}
\item \href{#method-Page-context}{\code{Page$context()}}
\item \href{#method-Page-drag_and_drop}{\code{Page$drag_and_drop()}}
\item \href{#method-Page-emulate_media}{\code{Page$emulate_media()}}
\item \href{#method-Page-evaluate}{\code{Page$evaluate()}}
\item \href{#method-Page-evaluate_handle}{\code{Page$evaluate_handle()}}
\item \href{#method-Page-expose_binding}{\code{Page$expose_binding()}}
\item \href{#method-Page-expose_function}{\code{Page$expose_function()}}
\item \href{#method-Page-frame}{\code{Page$frame()}}
\item \href{#method-Page-frame_locator}{\code{Page$frame_locator()}}
\item \href{#method-Page-frames}{\code{Page$frames()}}
\item \href{#method-Page-get_by_alt_text}{\code{Page$get_by_alt_text()}}
\item \href{#method-Page-get_by_label}{\code{Page$get_by_label()}}
\item \href{#method-Page-get_by_placeholder}{\code{Page$get_by_placeholder()}}
\item \href{#method-Page-get_by_role}{\code{Page$get_by_role()}}
\item \href{#method-Page-get_by_test_id}{\code{Page$get_by_test_id()}}
\item \href{#method-Page-get_by_text}{\code{Page$get_by_text()}}
\item \href{#method-Page-get_by_title}{\code{Page$get_by_title()}}
\item \href{#method-Page-go_back}{\code{Page$go_back()}}
\item \href{#method-Page-go_forward}{\code{Page$go_forward()}}
\item \href{#method-Page-goto}{\code{Page$goto()}}
\item \href{#method-Page-is_closed}{\code{Page$is_closed()}}
\item \href{#method-Page-locator}{\code{Page$locator()}}
\item \href{#method-Page-main_frame}{\code{Page$main_frame()}}
\item \href{#method-Page-opener}{\code{Page$opener()}}
\item \href{#method-Page-pause}{\code{Page$pause()}}
\item \href{#method-Page-pdf}{\code{Page$pdf()}}
\item \href{#method-Page-reload}{\code{Page$reload()}}
\item \href{#method-Page-route}{\code{Page$route()}}
\item \href{#method-Page-screenshot}{\code{Page$screenshot()}}
\item \href{#method-Page-set_content}{\code{Page$set_content()}}
\item \href{#method-Page-set_default_navigation_timeout}{\code{Page$set_default_navigation_timeout()}}
\item \href{#method-Page-set_default_timeout}{\code{Page$set_default_timeout()}}
\item \href{#method-Page-set_extra_h_t_t_p_headers}{\code{Page$set_extra_h_t_t_p_headers()}}
\item \href{#method-Page-set_viewport_size}{\code{Page$set_viewport_size()}}
\item \href{#method-Page-title}{\code{Page$title()}}
\item \href{#method-Page-unroute}{\code{Page$unroute()}}
\item \href{#method-Page-unroute_all}{\code{Page$unroute_all()}}
\item \href{#method-Page-url}{\code{Page$url()}}
\item \href{#method-Page-video}{\code{Page$video()}}
\item \href{#method-Page-viewport_size}{\code{Page$viewport_size()}}
\item \href{#method-Page-wait_for_event}{\code{Page$wait_for_event()}}
\item \href{#method-Page-wait_for_function}{\code{Page$wait_for_function()}}
\item \href{#method-Page-wait_for_load_state}{\code{Page$wait_for_load_state()}}
\item \href{#method-Page-wait_for_request}{\code{Page$wait_for_request()}}
\item \href{#method-Page-wait_for_response}{\code{Page$wait_for_response()}}
\item \href{#method-Page-wait_for_u_r_l}{\code{Page$wait_for_u_r_l()}}
\item \href{#method-Page-workers}{\code{Page$workers()}}
\item \href{#method-Page-new}{\code{Page$new()}}
\item \href{#method-Page-clone}{\code{Page$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-add_locator_handler"></a>}}
\if{latex}{\out{\hypertarget{method-Page-add_locator_handler}{}}}
\subsection{Method \code{add_locator_handler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$add_locator_handler()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-remove_locator_handler"></a>}}
\if{latex}{\out{\hypertarget{method-Page-remove_locator_handler}{}}}
\subsection{Method \code{remove_locator_handler()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$remove_locator_handler()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-route_from_h_a_r"></a>}}
\if{latex}{\out{\hypertarget{method-Page-route_from_h_a_r}{}}}
\subsection{Method \code{route_from_h_a_r()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$route_from_h_a_r()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-add_init_script"></a>}}
\if{latex}{\out{\hypertarget{method-Page-add_init_script}{}}}
\subsection{Method \code{add_init_script()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$add_init_script(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-add_script_tag"></a>}}
\if{latex}{\out{\hypertarget{method-Page-add_script_tag}{}}}
\subsection{Method \code{add_script_tag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$add_script_tag(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-add_style_tag"></a>}}
\if{latex}{\out{\hypertarget{method-Page-add_style_tag}{}}}
\subsection{Method \code{add_style_tag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$add_style_tag(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-bring_to_front"></a>}}
\if{latex}{\out{\hypertarget{method-Page-bring_to_front}{}}}
\subsection{Method \code{bring_to_front()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$bring_to_front(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-close"></a>}}
\if{latex}{\out{\hypertarget{method-Page-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$close(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-content"></a>}}
\if{latex}{\out{\hypertarget{method-Page-content}{}}}
\subsection{Method \code{content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$content(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-context"></a>}}
\if{latex}{\out{\hypertarget{method-Page-context}{}}}
\subsection{Method \code{context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$context(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-drag_and_drop"></a>}}
\if{latex}{\out{\hypertarget{method-Page-drag_and_drop}{}}}
\subsection{Method \code{drag_and_drop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$drag_and_drop(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-emulate_media"></a>}}
\if{latex}{\out{\hypertarget{method-Page-emulate_media}{}}}
\subsection{Method \code{emulate_media()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$emulate_media(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-Page-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$evaluate(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-evaluate_handle"></a>}}
\if{latex}{\out{\hypertarget{method-Page-evaluate_handle}{}}}
\subsection{Method \code{evaluate_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$evaluate_handle(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-expose_binding"></a>}}
\if{latex}{\out{\hypertarget{method-Page-expose_binding}{}}}
\subsection{Method \code{expose_binding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$expose_binding(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-expose_function"></a>}}
\if{latex}{\out{\hypertarget{method-Page-expose_function}{}}}
\subsection{Method \code{expose_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$expose_function(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-frame"></a>}}
\if{latex}{\out{\hypertarget{method-Page-frame}{}}}
\subsection{Method \code{frame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$frame(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-frame_locator"></a>}}
\if{latex}{\out{\hypertarget{method-Page-frame_locator}{}}}
\subsection{Method \code{frame_locator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$frame_locator(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-frames"></a>}}
\if{latex}{\out{\hypertarget{method-Page-frames}{}}}
\subsection{Method \code{frames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$frames(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-get_by_alt_text"></a>}}
\if{latex}{\out{\hypertarget{method-Page-get_by_alt_text}{}}}
\subsection{Method \code{get_by_alt_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$get_by_alt_text(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-get_by_label"></a>}}
\if{latex}{\out{\hypertarget{method-Page-get_by_label}{}}}
\subsection{Method \code{get_by_label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$get_by_label(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-get_by_placeholder"></a>}}
\if{latex}{\out{\hypertarget{method-Page-get_by_placeholder}{}}}
\subsection{Method \code{get_by_placeholder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$get_by_placeholder(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-get_by_role"></a>}}
\if{latex}{\out{\hypertarget{method-Page-get_by_role}{}}}
\subsection{Method \code{get_by_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$get_by_role(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-get_by_test_id"></a>}}
\if{latex}{\out{\hypertarget{method-Page-get_by_test_id}{}}}
\subsection{Method \code{get_by_test_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$get_by_test_id(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-get_by_text"></a>}}
\if{latex}{\out{\hypertarget{method-Page-get_by_text}{}}}
\subsection{Method \code{get_by_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$get_by_text(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-get_by_title"></a>}}
\if{latex}{\out{\hypertarget{method-Page-get_by_title}{}}}
\subsection{Method \code{get_by_title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$get_by_title(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-go_back"></a>}}
\if{latex}{\out{\hypertarget{method-Page-go_back}{}}}
\subsection{Method \code{go_back()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$go_back(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-go_forward"></a>}}
\if{latex}{\out{\hypertarget{method-Page-go_forward}{}}}
\subsection{Method \code{go_forward()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$go_forward(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-goto"></a>}}
\if{latex}{\out{\hypertarget{method-Page-goto}{}}}
\subsection{Method \code{goto()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$goto(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-is_closed"></a>}}
\if{latex}{\out{\hypertarget{method-Page-is_closed}{}}}
\subsection{Method \code{is_closed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$is_closed(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-locator"></a>}}
\if{latex}{\out{\hypertarget{method-Page-locator}{}}}
\subsection{Method \code{locator()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$locator(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-main_frame"></a>}}
\if{latex}{\out{\hypertarget{method-Page-main_frame}{}}}
\subsection{Method \code{main_frame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$main_frame(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-opener"></a>}}
\if{latex}{\out{\hypertarget{method-Page-opener}{}}}
\subsection{Method \code{opener()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$opener(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-pause"></a>}}
\if{latex}{\out{\hypertarget{method-Page-pause}{}}}
\subsection{Method \code{pause()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$pause(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-Page-pdf}{}}}
\subsection{Method \code{pdf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$pdf(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-reload"></a>}}
\if{latex}{\out{\hypertarget{method-Page-reload}{}}}
\subsection{Method \code{reload()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$reload(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-route"></a>}}
\if{latex}{\out{\hypertarget{method-Page-route}{}}}
\subsection{Method \code{route()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$route(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-screenshot"></a>}}
\if{latex}{\out{\hypertarget{method-Page-screenshot}{}}}
\subsection{Method \code{screenshot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$screenshot(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-set_content"></a>}}
\if{latex}{\out{\hypertarget{method-Page-set_content}{}}}
\subsection{Method \code{set_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$set_content(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-set_default_navigation_timeout"></a>}}
\if{latex}{\out{\hypertarget{method-Page-set_default_navigation_timeout}{}}}
\subsection{Method \code{set_default_navigation_timeout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$set_default_navigation_timeout(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-set_default_timeout"></a>}}
\if{latex}{\out{\hypertarget{method-Page-set_default_timeout}{}}}
\subsection{Method \code{set_default_timeout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$set_default_timeout(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-set_extra_h_t_t_p_headers"></a>}}
\if{latex}{\out{\hypertarget{method-Page-set_extra_h_t_t_p_headers}{}}}
\subsection{Method \code{set_extra_h_t_t_p_headers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$set_extra_h_t_t_p_headers(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-set_viewport_size"></a>}}
\if{latex}{\out{\hypertarget{method-Page-set_viewport_size}{}}}
\subsection{Method \code{set_viewport_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$set_viewport_size(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-title"></a>}}
\if{latex}{\out{\hypertarget{method-Page-title}{}}}
\subsection{Method \code{title()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$title(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-unroute"></a>}}
\if{latex}{\out{\hypertarget{method-Page-unroute}{}}}
\subsection{Method \code{unroute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$unroute(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-unroute_all"></a>}}
\if{latex}{\out{\hypertarget{method-Page-unroute_all}{}}}
\subsection{Method \code{unroute_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$unroute_all(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-url"></a>}}
\if{latex}{\out{\hypertarget{method-Page-url}{}}}
\subsection{Method \code{url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$url(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-video"></a>}}
\if{latex}{\out{\hypertarget{method-Page-video}{}}}
\subsection{Method \code{video()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$video(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-viewport_size"></a>}}
\if{latex}{\out{\hypertarget{method-Page-viewport_size}{}}}
\subsection{Method \code{viewport_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$viewport_size(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-wait_for_event"></a>}}
\if{latex}{\out{\hypertarget{method-Page-wait_for_event}{}}}
\subsection{Method \code{wait_for_event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$wait_for_event(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-wait_for_function"></a>}}
\if{latex}{\out{\hypertarget{method-Page-wait_for_function}{}}}
\subsection{Method \code{wait_for_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$wait_for_function(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-wait_for_load_state"></a>}}
\if{latex}{\out{\hypertarget{method-Page-wait_for_load_state}{}}}
\subsection{Method \code{wait_for_load_state()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$wait_for_load_state(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-wait_for_request"></a>}}
\if{latex}{\out{\hypertarget{method-Page-wait_for_request}{}}}
\subsection{Method \code{wait_for_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$wait_for_request(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-wait_for_response"></a>}}
\if{latex}{\out{\hypertarget{method-Page-wait_for_response}{}}}
\subsection{Method \code{wait_for_response()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$wait_for_response(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-wait_for_u_r_l"></a>}}
\if{latex}{\out{\hypertarget{method-Page-wait_for_u_r_l}{}}}
\subsection{Method \code{wait_for_u_r_l()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$wait_for_u_r_l(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-workers"></a>}}
\if{latex}{\out{\hypertarget{method-Page-workers}{}}}
\subsection{Method \code{workers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$workers(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-new"></a>}}
\if{latex}{\out{\hypertarget{method-Page-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$new(parent, async = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Page-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
