% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_promise.R
\name{Promise}
\alias{Promise}
\title{Promise Class}
\description{
Promise Class

Promise Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Promise-then}{\code{Promise$then()}}
\item \href{#method-Promise-new}{\code{Promise$new()}}
\item \href{#method-Promise-clone}{\code{Promise$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Promise-then"></a>}}
\if{latex}{\out{\hypertarget{method-Promise-then}{}}}
\subsection{Method \code{then()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Promise$then(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Promise-new"></a>}}
\if{latex}{\out{\hypertarget{method-Promise-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Promise$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Promise-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Promise-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Promise$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
