% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_request.R
\name{Request}
\alias{Request}
\title{Request Class}
\description{
Request Class

Request Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Request-all_headers}{\code{Request$all_headers()}}
\item \href{#method-Request-failure}{\code{Request$failure()}}
\item \href{#method-Request-frame}{\code{Request$frame()}}
\item \href{#method-Request-header_value}{\code{Request$header_value()}}
\item \href{#method-Request-headers}{\code{Request$headers()}}
\item \href{#method-Request-headers_array}{\code{Request$headers_array()}}
\item \href{#method-Request-is_navigation_request}{\code{Request$is_navigation_request()}}
\item \href{#method-Request-method}{\code{Request$method()}}
\item \href{#method-Request-post_data}{\code{Request$post_data()}}
\item \href{#method-Request-post_data_buffer}{\code{Request$post_data_buffer()}}
\item \href{#method-Request-post_data_j_s_o_n}{\code{Request$post_data_j_s_o_n()}}
\item \href{#method-Request-redirected_from}{\code{Request$redirected_from()}}
\item \href{#method-Request-redirected_to}{\code{Request$redirected_to()}}
\item \href{#method-Request-resource_type}{\code{Request$resource_type()}}
\item \href{#method-Request-response}{\code{Request$response()}}
\item \href{#method-Request-service_worker}{\code{Request$service_worker()}}
\item \href{#method-Request-sizes}{\code{Request$sizes()}}
\item \href{#method-Request-timing}{\code{Request$timing()}}
\item \href{#method-Request-url}{\code{Request$url()}}
\item \href{#method-Request-new}{\code{Request$new()}}
\item \href{#method-Request-clone}{\code{Request$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-all_headers"></a>}}
\if{latex}{\out{\hypertarget{method-Request-all_headers}{}}}
\subsection{Method \code{all_headers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$all_headers(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-failure"></a>}}
\if{latex}{\out{\hypertarget{method-Request-failure}{}}}
\subsection{Method \code{failure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$failure(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-frame"></a>}}
\if{latex}{\out{\hypertarget{method-Request-frame}{}}}
\subsection{Method \code{frame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$frame(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-header_value"></a>}}
\if{latex}{\out{\hypertarget{method-Request-header_value}{}}}
\subsection{Method \code{header_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$header_value(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-headers"></a>}}
\if{latex}{\out{\hypertarget{method-Request-headers}{}}}
\subsection{Method \code{headers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$headers(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-headers_array"></a>}}
\if{latex}{\out{\hypertarget{method-Request-headers_array}{}}}
\subsection{Method \code{headers_array()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$headers_array(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-is_navigation_request"></a>}}
\if{latex}{\out{\hypertarget{method-Request-is_navigation_request}{}}}
\subsection{Method \code{is_navigation_request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$is_navigation_request(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-method"></a>}}
\if{latex}{\out{\hypertarget{method-Request-method}{}}}
\subsection{Method \code{method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$method(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-post_data"></a>}}
\if{latex}{\out{\hypertarget{method-Request-post_data}{}}}
\subsection{Method \code{post_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$post_data(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-post_data_buffer"></a>}}
\if{latex}{\out{\hypertarget{method-Request-post_data_buffer}{}}}
\subsection{Method \code{post_data_buffer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$post_data_buffer(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-post_data_j_s_o_n"></a>}}
\if{latex}{\out{\hypertarget{method-Request-post_data_j_s_o_n}{}}}
\subsection{Method \code{post_data_j_s_o_n()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$post_data_j_s_o_n(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-redirected_from"></a>}}
\if{latex}{\out{\hypertarget{method-Request-redirected_from}{}}}
\subsection{Method \code{redirected_from()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$redirected_from(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-redirected_to"></a>}}
\if{latex}{\out{\hypertarget{method-Request-redirected_to}{}}}
\subsection{Method \code{redirected_to()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$redirected_to(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-resource_type"></a>}}
\if{latex}{\out{\hypertarget{method-Request-resource_type}{}}}
\subsection{Method \code{resource_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$resource_type(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-response"></a>}}
\if{latex}{\out{\hypertarget{method-Request-response}{}}}
\subsection{Method \code{response()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$response(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-service_worker"></a>}}
\if{latex}{\out{\hypertarget{method-Request-service_worker}{}}}
\subsection{Method \code{service_worker()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$service_worker(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-sizes"></a>}}
\if{latex}{\out{\hypertarget{method-Request-sizes}{}}}
\subsection{Method \code{sizes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$sizes(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-timing"></a>}}
\if{latex}{\out{\hypertarget{method-Request-timing}{}}}
\subsection{Method \code{timing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$timing(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-url"></a>}}
\if{latex}{\out{\hypertarget{method-Request-url}{}}}
\subsection{Method \code{url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$url(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-new"></a>}}
\if{latex}{\out{\hypertarget{method-Request-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Request-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Request-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Request$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
