% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_response.R
\name{Response}
\alias{Response}
\title{Response Class}
\description{
Response Class

Response Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Response-frame}{\code{Response$frame()}}
\item \href{#method-Response-from_service_worker}{\code{Response$from_service_worker()}}
\item \href{#method-Response-header_value}{\code{Response$header_value()}}
\item \href{#method-Response-header_values}{\code{Response$header_values()}}
\item \href{#method-Response-request}{\code{Response$request()}}
\item \href{#method-Response-all_headers}{\code{Response$all_headers()}}
\item \href{#method-Response-body}{\code{Response$body()}}
\item \href{#method-Response-finished}{\code{Response$finished()}}
\item \href{#method-Response-headers}{\code{Response$headers()}}
\item \href{#method-Response-headers_array}{\code{Response$headers_array()}}
\item \href{#method-Response-json}{\code{Response$json()}}
\item \href{#method-Response-ok}{\code{Response$ok()}}
\item \href{#method-Response-security_details}{\code{Response$security_details()}}
\item \href{#method-Response-server_addr}{\code{Response$server_addr()}}
\item \href{#method-Response-status}{\code{Response$status()}}
\item \href{#method-Response-status_text}{\code{Response$status_text()}}
\item \href{#method-Response-text}{\code{Response$text()}}
\item \href{#method-Response-url}{\code{Response$url()}}
\item \href{#method-Response-new}{\code{Response$new()}}
\item \href{#method-Response-clone}{\code{Response$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-frame"></a>}}
\if{latex}{\out{\hypertarget{method-Response-frame}{}}}
\subsection{Method \code{frame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$frame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-from_service_worker"></a>}}
\if{latex}{\out{\hypertarget{method-Response-from_service_worker}{}}}
\subsection{Method \code{from_service_worker()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$from_service_worker()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-header_value"></a>}}
\if{latex}{\out{\hypertarget{method-Response-header_value}{}}}
\subsection{Method \code{header_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$header_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-header_values"></a>}}
\if{latex}{\out{\hypertarget{method-Response-header_values}{}}}
\subsection{Method \code{header_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$header_values()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-request"></a>}}
\if{latex}{\out{\hypertarget{method-Response-request}{}}}
\subsection{Method \code{request()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$request()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-all_headers"></a>}}
\if{latex}{\out{\hypertarget{method-Response-all_headers}{}}}
\subsection{Method \code{all_headers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$all_headers(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-body"></a>}}
\if{latex}{\out{\hypertarget{method-Response-body}{}}}
\subsection{Method \code{body()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$body(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-finished"></a>}}
\if{latex}{\out{\hypertarget{method-Response-finished}{}}}
\subsection{Method \code{finished()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$finished(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-headers"></a>}}
\if{latex}{\out{\hypertarget{method-Response-headers}{}}}
\subsection{Method \code{headers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$headers(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-headers_array"></a>}}
\if{latex}{\out{\hypertarget{method-Response-headers_array}{}}}
\subsection{Method \code{headers_array()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$headers_array(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-json"></a>}}
\if{latex}{\out{\hypertarget{method-Response-json}{}}}
\subsection{Method \code{json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$json(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-ok"></a>}}
\if{latex}{\out{\hypertarget{method-Response-ok}{}}}
\subsection{Method \code{ok()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$ok(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-security_details"></a>}}
\if{latex}{\out{\hypertarget{method-Response-security_details}{}}}
\subsection{Method \code{security_details()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$security_details(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-server_addr"></a>}}
\if{latex}{\out{\hypertarget{method-Response-server_addr}{}}}
\subsection{Method \code{server_addr()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$server_addr(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-status"></a>}}
\if{latex}{\out{\hypertarget{method-Response-status}{}}}
\subsection{Method \code{status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$status(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-status_text"></a>}}
\if{latex}{\out{\hypertarget{method-Response-status_text}{}}}
\subsection{Method \code{status_text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$status_text(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-text"></a>}}
\if{latex}{\out{\hypertarget{method-Response-text}{}}}
\subsection{Method \code{text()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$text(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-url"></a>}}
\if{latex}{\out{\hypertarget{method-Response-url}{}}}
\subsection{Method \code{url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$url(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-new"></a>}}
\if{latex}{\out{\hypertarget{method-Response-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Response-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Response-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Response$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
