% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplaywright.R
\name{rplaywright}
\alias{rplaywright}
\title{An R wrapper for playwright}
\description{
These are R wrapper for playwright
}
\references{
http://github.com/erikaris/rselenium
}
\author{
Erika Mukhlisina Siregar
}
