% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_browser_close}
\alias{rplaywright_browser_close}
\title{Close existing browser instance}
\usage{
rplaywright_browser_close(browser)
}
\arguments{
\item{browser}{Browser instance to close}
}
\description{
Close existing browser instance
}
\examples{
rplaywright::rplaywright_browser_close(browser)

}
