% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_browser_new}
\alias{rplaywright_browser_new}
\title{Launch new browser instance}
\usage{
rplaywright_browser_new(type = "chromium", startServer = T)
}
\arguments{
\item{type}{Type of browser to launch \code{chromium, firefox, webkit, NULL}}
}
\description{
Launch new browser instance
}
\examples{
browser <- rplaywright::rplaywright_browser_new(type='chromium')

}
