% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_context_close}
\alias{rplaywright_context_close}
\title{Close existing context instance}
\usage{
rplaywright_context_close(context)
}
\arguments{
\item{context}{Context instance to close}
}
\description{
Close existing context instance
}
\examples{
rplaywright::rplaywright_context_close(context)

}
