% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_context_new}
\alias{rplaywright_context_new}
\title{Create new browser context}
\usage{
rplaywright_context_new(browser, options = list())
}
\arguments{
\item{browser}{Browser instance}

\item{options}{Context options as defined in \url{https://playwright.dev/docs/api/class-browser#browser-new-context}}
}
\description{
Create new browser context
}
\examples{
context <- rplaywright::rplaywright_context_new(browser, options = list())

}
