% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_npm_install}
\alias{rplaywright_npm_install}
\title{a wrapper for nodejs playwright functionalities}
\usage{
rplaywright_npm_install(force = FALSE)
}
\description{
here we will put more details
}
