% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_page_close}
\alias{rplaywright_page_close}
\title{Close existing page instance}
\usage{
rplaywright_page_close(page)
}
\arguments{
\item{page}{Page instance to close}
}
\description{
Close existing page instance
}
\examples{
rplaywright::rplaywright_page_close(page)

}
