% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_page_getbylabel}
\alias{rplaywright_page_getbylabel}
\title{Locating input by label}
\usage{
rplaywright_page_getbylabel(page, args = list(), actions = list())
}
\arguments{
\item{page}{Page instance}

\item{args}{List of as defined in \url{https://playwright.dev/docs/api/class-page#page-get-by-label}}

\item{actions}{List of actions as defined in \url{https://playwright.dev/docs/api/class-locator}, such as \code{fill} (\url{https://playwright.dev/docs/api/class-locator#locator-fill})}
}
\description{
Allows locating input elements by the text of the associated \if{html}{\out{<label>}} or aria-labelledby element, or by the aria-label attribute.
Optinally, do actions refered in \url{https://playwright.dev/docs/api/class-locator}, such as \code{fill} (\url{https://playwright.dev/docs/api/class-locator#locator-fill})
}
\details{
\code{
<input aria-label="Username" />
<label for="password-input">Password:</label>
<input id="password-input" />
}
}
\examples{
rplaywright::rplaywright_page_getbylabel(page, args = list(text = "Username"), actions = list(fill = list("john")))
rplaywright::rplaywright_page_getbylabel(page, args = list(text = "Password"), actions = list(fill = list("password")))

}
