% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_page_getbyrole}
\alias{rplaywright_page_getbyrole}
\title{Locating input by Role}
\usage{
rplaywright_page_getbyrole(page, args = list(), actions = list())
}
\arguments{
\item{page}{Page instance}

\item{args}{List of as defined in \url{https://playwright.dev/docs/api/class-page#page-get-by-role}}

\item{actions}{List of actions as defined in \url{https://playwright.dev/docs/api/class-locator}, such as \code{click} (\url{https://playwright.dev/docs/api/class-locator#locator-click})}
}
\description{
Allows locating elements by their ARIA role, ARIA attributes and accessible name.
Optinally, do actions refered in \url{https://playwright.dev/docs/api/class-locator}, such as \code{fill} (\url{https://playwright.dev/docs/api/class-locator#locator-fill})
}
\details{
\code{
<h3>Sign up</h3>
<label>
  <input type="checkbox" /> Subscribe
</label>
<br/>
<button>Submit</button>
}
}
\examples{
rplaywright::rplaywright_page_getbyrole(page, args = list(role = "checkbox", options = list(text = "Subscribe")), actions = list(check = list()))
rplaywright::rplaywright_page_getbyrole(page, args = list(role = "button", options = list(text = "Submit")), actions = list(click = list()))

}
