% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_page_goto}
\alias{rplaywright_page_goto}
\title{Open \code{url} in the existing page}
\usage{
rplaywright_page_goto(page, url, async = F)
}
\arguments{
\item{page}{Page instance}

\item{url}{Url to open}

\item{async}{If {TRUE}, this function will not wait until new page loaded}
}
\description{
Open \code{url} in the existing page
}
\examples{
rplaywright::rplaywright_page_goto(page, url = 'https://github.com/erikaris/rplaywright', async = F)

}
