% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_page_new}
\alias{rplaywright_page_new}
\title{Open new page, with \code{url} or with \code{content} if provide. Otherwise it will open blank page.}
\usage{
rplaywright_page_new(context, url, content, async = F)
}
\arguments{
\item{context}{Context instance}

\item{url}{Url to open}

\item{content}{Content to display}

\item{async}{If {TRUE}, this function will not wait until new page loaded}
}
\description{
Open new page, with \code{url} or with \code{content} if provide. Otherwise it will open blank page.
}
\examples{
page <- rplaywright::rplaywright_page_new(context, url = 'https://github.com/erikaris/rplaywright', async = F)

}
