% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_page_screenshot}
\alias{rplaywright_page_screenshot}
\title{Take screenshot of existing page}
\usage{
rplaywright_page_screenshot(page, path)
}
\arguments{
\item{page}{Page instance}

\item{path}{Path for screenshot to save, optinal}
}
\description{
Take screenshot of existing page
}
\examples{
rplaywright::rplaywright_page_screenshot(page, path = './ss1.png')

}
