% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_page_scrollto}
\alias{rplaywright_page_scrollto}
\title{Scroll page to spesific location}
\usage{
rplaywright_page_scrollto(page, args = list())
}
\arguments{
\item{page}{Page instance}

\item{args}{One of top or left}
}
\description{
It is not part of playwright API
}
\details{
\code{
window.scrollTo({
  behavior: "smooth",
  top: 10,
  left: 10,
})
}
}
\examples{
rplaywright::rplaywright_page_scrollto(page, args = list(top = 10, left = 10))

}
