% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_page_scrolltoright}
\alias{rplaywright_page_scrolltoright}
\title{Scroll page to most right}
\usage{
rplaywright_page_scrolltoright(page)
}
\arguments{
\item{page}{Page instance}
}
\description{
It is not part of playwright API
}
\details{
\code{
window.scrollTo({
  behavior: "smooth",
  top: document.body.scrollWidth,
})
}
}
\examples{
rplaywright::rplaywright_page_scrolltoright(page)

}
