% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_page_set_content}
\alias{rplaywright_page_set_content}
\title{Display content in the existing page}
\usage{
rplaywright_page_set_content(page, content, async = F)
}
\arguments{
\item{page}{Page instance}

\item{content}{Content to display}

\item{async}{If {TRUE}, this function will not wait until new page loaded}
}
\description{
Display content in the existing page
}
\examples{
rplaywright::rplaywright_page_set_content(page, content = '<h2>Title</h2>', async = F)

}
