% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_page_waitforresponse}
\alias{rplaywright_page_waitforresponse}
\title{Returns the matched response}
\usage{
rplaywright_page_waitforresponse(page, async = F, args = list())
}
\arguments{
\item{page}{Page instance}

\item{async}{If {TRUE}, this function will not wait until new page loaded}

\item{args}{List of as defined in \url{https://playwright.dev/docs/api/class-page#page-wait-for-timeout}}
}
\value{
future if async or httr if not async
}
\description{
if async = T, the value must be resolved with future::value()
}
\examples{
rplaywright::rplaywright_page_waitforresponse(page1, args = list(
  timeout = 5000,
  filter = list(or = list(
    list(url = "SearchTimeline"),
    list(url = "TweetDetail")
  ))
))

}
