% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_page_waitfortimeout}
\alias{rplaywright_page_waitfortimeout}
\title{Waits for the given timeout in milliseconds.}
\usage{
rplaywright_page_waitfortimeout(page, args = list())
}
\arguments{
\item{page}{Page instance}

\item{args}{List of as defined in \url{https://playwright.dev/docs/api/class-page#page-wait-for-timeout}}
}
\description{
Waits for the given timeout in milliseconds.
}
\examples{
rplaywright::rplaywright_page_waitfortimeout(page, args = list(timeout = 1000))

}
