% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_start_server}
\alias{rplaywright_start_server}
\title{function for starting the headless browser server.}
\usage{
rplaywright_start_server(force = FALSE)
}
\arguments{
\item{force}{}
}
\description{
function for starting the headless browser server.
}
