% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_stop_server}
\alias{rplaywright_stop_server}
\title{Stop rplaywright server}
\usage{
rplaywright_stop_server()
}
\description{
Stop rplaywright server
}
\examples{
rplaywright::rplaywright_stop_server()

}
