% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_rplaywright.R
\name{start_server}
\alias{start_server}
\title{Start rplaywright server}
\usage{
start_server(host = "127.0.0.1", port = 3000, restart = F)
}
\arguments{
\item{host}{Rplaywright server host. Default to 127.0.0.1}

\item{port}{Rplaywright server port. Default to 3000}
}
\description{
Start rplaywright server
}
\examples{
\dontrun{
rplaywright::start_server()
}

}
