% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_video.R
\name{Video}
\alias{Video}
\title{Video Class}
\description{
Video Class

Video Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Video-delete}{\code{Video$delete()}}
\item \href{#method-Video-path}{\code{Video$path()}}
\item \href{#method-Video-save_as}{\code{Video$save_as()}}
\item \href{#method-Video-new}{\code{Video$new()}}
\item \href{#method-Video-clone}{\code{Video$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Video-delete"></a>}}
\if{latex}{\out{\hypertarget{method-Video-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Video$delete(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Video-path"></a>}}
\if{latex}{\out{\hypertarget{method-Video-path}{}}}
\subsection{Method \code{path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Video$path(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Video-save_as"></a>}}
\if{latex}{\out{\hypertarget{method-Video-save_as}{}}}
\subsection{Method \code{save_as()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Video$save_as(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Video-new"></a>}}
\if{latex}{\out{\hypertarget{method-Video-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Video$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Video-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Video-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Video$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
