% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_worker.R
\name{Worker}
\alias{Worker}
\title{Worker Class}
\description{
Worker Class

Worker Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Worker-evaluate_handle}{\code{Worker$evaluate_handle()}}
\item \href{#method-Worker-evaluate}{\code{Worker$evaluate()}}
\item \href{#method-Worker-url}{\code{Worker$url()}}
\item \href{#method-Worker-new}{\code{Worker$new()}}
\item \href{#method-Worker-clone}{\code{Worker$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Worker-evaluate_handle"></a>}}
\if{latex}{\out{\hypertarget{method-Worker-evaluate_handle}{}}}
\subsection{Method \code{evaluate_handle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Worker$evaluate_handle()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Worker-evaluate"></a>}}
\if{latex}{\out{\hypertarget{method-Worker-evaluate}{}}}
\subsection{Method \code{evaluate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Worker$evaluate(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Worker-url"></a>}}
\if{latex}{\out{\hypertarget{method-Worker-url}{}}}
\subsection{Method \code{url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Worker$url(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Worker-new"></a>}}
\if{latex}{\out{\hypertarget{method-Worker-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Worker$new(parent)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Worker-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Worker-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Worker$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
