% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_rplaywright.R
\name{new_chromium}
\alias{new_chromium}
\title{Launch new chromium instance}
\usage{
new_chromium(
  start_server = T,
  serverOptions = list(host = "127.0.0.1", port = 3000)
)
}
\value{
An object of class \code{Browser}
}
\description{
Launch new chromium instance
}
\examples{
\dontrun{
browser <- rplaywright::new_chromium()
}

}
