% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_npm_install}
\alias{rplaywright_npm_install}
\title{Install nodejs dependencies}
\usage{
rplaywright_npm_install(force = FALSE)
}
\arguments{
\item{force}{Install dependencies without prompt}
}
\description{
Install nodejs dependencies
}
\examples{
rplaywright::rplaywright_npm_install(force=T)

}
