% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infra.R
\name{rplaywright_start_server}
\alias{rplaywright_start_server}
\title{Start rplaywright server}
\usage{
rplaywright_start_server()
}
\description{
Start rplaywright server
}
\examples{
rplaywright::rplaywright_start_server()

}
